+++ src/core_modules/capture-core/components/WidgetEventSchedule/ScheduleButtons/scheduleButtons.types.js
@@ -1,0 +1,7 @@
+// @flow
+
+export type Props = {|
+    onSchedule: () => Promise<void>,
+    onCancel: () => void,
+    ...CssClasses,
+|};